\name{QGglmm-package}
\alias{QGglmm-package}
\alias{QGglmm}
\docType{package}
\title{
Quantitative Genetics parameters for Generalised Linear Mixed Models
}
\description{
This package gives the values on the observed scale for several quantitative genetics parameter. If a fitness function is assumed or measured, it also predicts the evolutionary response to selection on the observed scale.
}
\details{
\tabular{ll}{
Package: \tab QGglmm\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-03-20\cr
License: \tab GPL-2\cr
}

\author{ Pierre de Villemereuil & Michael B. Morrissey.
Maintainer: Pierre de Villemereuil <bonamy@horus.ens.fr>
}

\references{
\strong{REF!!}
}
